Create MouthControl.py
This scripts uses MouhtControl. It also use 3 different positions for the jaw so that the jaw moves a bit open when speaking, and it closes the mouth fully when it stops speaking. 
Warning!!!!!! Since this script doesn't detach the jaw, it may burn the servo if it closes the jaw to much. !!!!!!