Create solar_display_v1
This solar program reads data from an OutBack Mate and displays it on a computer running MRL. The OutBack Mate is connected to 3 solar chargers and 2 DC to AC inverters. The OutBack Mate streams data at 19200 baud all the time through a serial port. The data stream contains lots of information about what is going on with your solar power system. All of the equipment monitored is from OutBack Power and connects through there own bus.